website/requirements.txt
alembic==1.10.2
apturl==0.5.2
async-timeout==4.0.2
bcrypt==3.2.0
blinker==1.5
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
exceptiongroup==1.1.1
fabric==2.6.0
fasteners==0.14.1
Flask==2.2.3
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
flask-redis==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
future==0.18.2
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.12.0
iniconfig==2.0.0
invoke==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==3.5.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
packaging==23.1
paramiko==2.9.3
pathlib2==2.3.7.post1
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==7.3.1
pytest-flask==1.2.0
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.5
reportlab==3.6.8
requests==2.25.1
rich==13.3.5
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.7
ssh-import-id==5.11
systemd-python==234
tomli==2.0.1
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uWSGI==2.0.21
wadllib==1.3.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
